Update main.py
Here the code was written for adding two numbers, but it had an error in the logic. Instead of adding two numbers it also added an extra 1 and then returned the value to the function.
So, it needed just to erase the extra 1 and just return the value of sum of 2 numbers.